requirements.txt
vllm
litellm
datasets
pyyaml
pandas
antlr4-python3-runtime
sympy
latex2sympy2_extended[antlr4_13_2]
math-verify[antlr4_13_2]